Improved FIND REFERENCE SQL command to retrieve multiple RID references

+++ core/src/main/java/com/orientechnologies/orient/core/sql/OCommandExecutorSQLDelete.java
@@ -147,10 +147,13 @@ public class OCommandExecutorSQLDelete extends OCommandExecutorSQLAbstract imple
 	 */
 	public boolean result(final Object iRecord) {
 		final ORecordAbstract<?> record = (ORecordAbstract<?>) iRecord;
-		record.delete();
 
-		recordCount++;
-		return true;
+		if (record.getIdentity().isValid()) {
+			record.delete();
+			recordCount++;
+			return true;
+		}
+		return false;
 	}
 
 	public String getSyntax() {
